test: split oversized Antigravity suites
